+++ src/modules/product/product.service.ts
@@ -1,7 +1,13 @@
 import { ProductModel } from "./product.model.js";
 import { Sort } from "../../common/enums/sort.enum.js";
 import type { IFilterCommon } from "../../common/interfaces/filter.interface.js";
-import type { Product, ProductUpdate } from "./product.schema.js";
+import {
+  ProductUpdateValidation,
+  ProductValidation,
+  type Product,
+  type ProductUpdate,
+} from "./product.schema.js";
+import mongoose from "mongoose";
 
 interface IProducts extends IFilterCommon {
   productCategory?: string;
@@ -13,7 +19,7 @@ export class ProductService {
     const {
       page = 1,
       limit = 10,
-      query = "",
+      query,
       sort = Sort.DESC,
       column = "createdAt",
     } = filter;
@@ -22,7 +28,7 @@ export class ProductService {
 
     const queryObj: any = {};
 
-    if (query !== "") {
+    if (query) {
       queryObj.$or = [
         { name: { $regex: query, $options: "i" } },
         { slug: { $regex: query, $options: "i" } },
@@ -42,11 +48,9 @@ export class ProductService {
       baseQuery.populate("product_category_id");
     }
 
-    console.log(queryObj);
-
     const [data, total] = await Promise.all([
       baseQuery,
-      ProductModel.countDocuments(queryObj),
+      ProductModel.countDocuments(),
     ]);
 
     return {
@@ -56,18 +60,61 @@ export class ProductService {
   }
 
   async findById(id: string | number) {
+    if (!id) {
+      throw new Error("ID không được bỏ trống");
+    }
+
+    if (!mongoose.Types.ObjectId.isValid(id as string)) {
+      throw new Error("ID phải có kiểu dữ liệu ObjectId");
+    }
+
     return await ProductModel.findById(id).populate("product_category_id");
   }
 
   async create(data: Product) {
-    return ProductModel.create(data);
+    await ProductValidation.parseAsync(data);
+
+    const resultCreate = await ProductModel.create(data);
+
+    if (!resultCreate) {
+      throw new Error("Tạo sản phẩm thất bại");
+    }
+
+    return resultCreate;
   }
 
   async remove(id: string) {
-    return ProductModel.findByIdAndDelete(id);
+    const resultDetail = await this.findById(id);
+    if (!resultDetail) {
+      throw new Error("Sản phẩm không tồn tại");
+    }
+
+    const resultDelete = await ProductModel.findByIdAndDelete(id);
+
+    if (!resultDelete) {
+      throw new Error("Xoa san pham that bai");
+    }
+
+    return resultDelete;
   }
 
   async update(id: string, data: ProductUpdate) {
-    return ProductModel.findByIdAndUpdate(id, data, { new: true });
+    const resultDetail = await this.findById(id);
+
+    if (!resultDetail) {
+      throw new Error("Sản phẩm không tồn tại");
+    }
+
+    await ProductUpdateValidation.parseAsync(data);
+
+    const resultUpdate = await ProductModel.findByIdAndUpdate(id, data, {
+      new: true,
+    });
+
+    if (!resultUpdate) {
+      throw new Error("Cập nhật sản phẩm thất bại");
+    }
+
+    return resultUpdate;
   }
 }
